Isolate scene detection from the engine facade
Scene detection is self-contained enough to leave the monolithic engine without changing the public import path. The new module preserves mcp_video.engine.detect_scenes while moving parsing and subprocess handling behind a narrower boundary.

Constraint: Keep client/server imports stable while continuing the approved engine decomposition plan.

Rejected: Extract larger analytics group with waveform and quality comparison | wider diff would raise conflict risk and slow review.

Confidence: high

Scope-risk: narrow

Directive: Keep engine.py as a compatibility facade for existing public imports until call sites are migrated intentionally.

Tested: ruff check mcp_video/engine.py mcp_video/engine_detect_scenes.py; ruff format --check mcp_video/engine.py mcp_video/engine_detect_scenes.py; python3 -m pytest tests/test_engine_advanced.py -k 'detect_scenes' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'detect_scenes' tests/test_server.py::TestVideoDetectScenesTool tests/test_red_team.py -k 'detect_scenes' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
